git-svn-id: file:///srv/local/data/cleanRepo/trunk@30950 a0517364-e873-4167-b1c8-416daa5a6b91

+++ playgrounds/johannes/src/main/java/playground/johannes/gsv/synPop/mid/sim/PersonLau2Inhabitants.java
@@ -87,7 +87,7 @@ public class PersonLau2Inhabitants implements Hamiltonian {
 		if(zone == null) {
 			return null;
 		} else {
-			Integer inhabs = (Integer)zone.getAttribute().get(LandUseData.POPULATION_KEY);
+			Integer inhabs = ((Double)zone.getAttribute().get(LandUseData.POPULATION_KEY)).intValue();
 			inhabitants.put(home, inhabs);
 			return inhabs;
 		}
